+++ src/components/Calculator.js
@@ -6,35 +6,11 @@ import {
   useCallback,
 } from 'react';
 import '../styles/Calculator.css';
+import { computeExpression, hasInput, computeNextOperand } from '../utils';
 import Expression from './Expression';
 import Digits from './Digits';
 import Modifier from './Modifier';
 import Operations from './Operations';
-
-const computeExpression = ({ firstOperand, secondOperand, operation }) => {
-  if (operation === '/') {
-    return Math.floor(firstOperand / secondOperand);
-  }
-  if (operation === 'X') {
-    return firstOperand * secondOperand;
-  }
-  if (operation === '-') {
-    return firstOperand - secondOperand;
-  }
-  if (operation === '+') {
-    return firstOperand + secondOperand;
-  }
-};
-
-const hasInput = ({ firstOperand, secondOperand, operation }) =>
-  firstOperand !== 0 || secondOperand !== -1 || operation !== null;
-
-const computeNextOperand = (currentOperand, digit) => {
-  currentOperand = String(currentOperand);
-  return currentOperand.length >= 3
-    ? Number(`${currentOperand.slice(0, -1)}${digit}`)
-    : Number(currentOperand + digit);
-};
 
 const Calculator = () => {
   const [state, setState] = useState({
